requirements.txt
albumentations==1.4.11
detectron2==0.6
easydict==1.13
fvcore==0.1.5.post20221221
imageio==2.34.1
ipython==8.12.3
lightning==2.2.5
matplotlib==3.7.2
MedPy==0.5.1
mmcv==2.2.0
mmengine==0.10.4
MultiScaleDeformableAttention==1.0
numpy==1.23.5
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==2.0.3
Pillow==9.4.0
pycm==4.0
pytorch_lightning==2.2.5
PyYAML==6.0.1
setproctitle==1.3.3
setuptools==60.2.0
tabulate==0.9.0
torch==2.1.0
torchinfo==1.8.0
torchmetrics==1.4.0.post0
torchvision==0.16.0
tqdm==4.65.2
transformers==4.41.2
